Full on refactor
