+++ LethalProgression/LP_NetworkManager.cs
@@ -1,7 +1,4 @@
 ﻿using HarmonyLib;
-using System;
-using System.Collections.Generic;
-using System.Text;
 using Unity.Netcode;
 using UnityEngine;
 using Object = UnityEngine.Object;
@@ -11,12 +8,16 @@ namespace LethalProgression
     [HarmonyPatch]
     internal class LP_NetworkManager
     {
+        public static LC_XP xpInstance;
+
         [HarmonyPostfix]
-        [HarmonyPatch(typeof(GameNetworkManager), "Start")]
+        [HarmonyPatch(typeof(GameNetworkManager), nameof(GameNetworkManager.Start))]
         public static void Init()
         {
             if (xpNetworkObject != null)
+            {
                 return;
+            }
 
             xpNetworkObject = (GameObject)LethalPlugin.skillBundle.LoadAsset("LP_XPHandler");
             xpNetworkObject.AddComponent<LC_XP>();
@@ -26,18 +27,17 @@ namespace LethalProgression
         public static GameObject xpNetworkObject;
 
         [HarmonyPostfix]
-        [HarmonyPatch(typeof(StartOfRound), "Awake")]
+        [HarmonyPatch(typeof(StartOfRound), nameof(StartOfRound.Awake))]
         static void SpawnNetworkHandler()
         {
-            if (NetworkManager.Singleton.IsHost || NetworkManager.Singleton.IsServer)
+            if (!NetworkManager.Singleton.IsHost && !NetworkManager.Singleton.IsServer)
             {
-                var networkHandlerHost = Object.Instantiate(xpNetworkObject, Vector3.zero, Quaternion.identity);
-                networkHandlerHost.GetComponent<NetworkObject>().Spawn();
-                xpInstance = networkHandlerHost.GetComponent<LC_XP>();
-                LethalPlugin.Log.LogInfo("XPHandler Initialized.");
+                return;
             }
+            var networkHandlerHost = Object.Instantiate(xpNetworkObject, Vector3.zero, Quaternion.identity);
+            networkHandlerHost.GetComponent<NetworkObject>().Spawn();
+            xpInstance = networkHandlerHost.GetComponent<LC_XP>();
+            LethalPlugin.Log.LogInfo("XPHandler Initialized.");
         }
-
-        public static LC_XP xpInstance;
     }
 }
