+++ LethalProgression/Plugin.cs
@@ -1,27 +1,17 @@
-﻿using BepInEx;
-using UnityEngine;
-using HarmonyLib;
-using BepInEx.Configuration;
-using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using BepInEx.Logging;
 using System.Reflection;
-using System.IO;
-using UnityEngine.SceneManagement;
+using BepInEx;
 using BepInEx.Bootstrap;
-using LethalProgression.GUI;
-using LethalProgression.Skills;
-using LethalProgression.Patches;
+using BepInEx.Configuration;
+using BepInEx.Logging;
+using HarmonyLib;
 using LethalProgression.Config;
-using Unity.Netcode;
 using UnityEngine;
 
 namespace LethalProgression
 {
-    [BepInPlugin("Stoneman.LethalProgression", "Lethal Progression", "1.4.0")]
+    [BepInPlugin(modGUID, modName, modVersion)]
     internal class LethalPlugin : BaseUnityPlugin
     {
         private const string modGUID = "Stoneman.LethalProgression";
@@ -39,7 +29,7 @@ namespace LethalProgression
         {
             Instance = this;
 
-            var harmony = new Harmony(modGUID);
+            Harmony harmony = new Harmony(modGUID);
             harmony.PatchAll(Assembly.GetExecutingAssembly());
 
             skillBundle = AssetBundle.LoadFromMemory(LethalProgression.Properties.Resources.skillmenu);
@@ -95,11 +85,7 @@ namespace LethalProgression
 
         public void BindConfig<T>(string section, string key, T defaultValue, string description = "")
         {
-            Config.Bind(section,
-                key,
-                defaultValue,
-                description
-            );
+            Config.Bind(section, key, defaultValue, description);
         }
 
         public IDictionary<string, string> GetAllConfigEntries()
