+++ LethalProgression/Saving/SaveManager.cs
@@ -1,8 +1,6 @@
-﻿using System.Collections.Generic;
-using UnityEngine;
+﻿using System.IO;
 using Newtonsoft.Json;
-using System.IO;
-using Steamworks;
+using UnityEngine;
 
 namespace LethalProgression.Saving
 {
@@ -13,7 +11,7 @@ namespace LethalProgression.Saving
         {
             saveFileSlot = GameNetworkManager.Instance.saveFileNum;
 
-            LethalPlugin.Log.LogInfo("Saving to slot " + saveFileSlot + 1);
+            LethalPlugin.Log.LogInfo($"Saving to slot {saveFileSlot + 1}");
 
             // If file doesn't exist, create it
             if (!Directory.Exists(GetSavePath()))
@@ -21,14 +19,14 @@ namespace LethalProgression.Saving
                 Directory.CreateDirectory(GetSavePath());
             }
 
-            File.WriteAllText(GetSavePath() + steamid + ".json", data);
+            File.WriteAllText($"{GetSavePath()}{steamid}.json", data);
         }
 
         public static void SaveShared(int xp, int level, int quota)
         {
             saveFileSlot = GameNetworkManager.Instance.saveFileNum;
 
-            LethalPlugin.Log.LogInfo("Saving to slot " + saveFileSlot + 1);
+            LethalPlugin.Log.LogInfo($"Saving to slot {saveFileSlot + 1}");
 
             // If file doesn't exist, create it
             if (!Directory.Exists(GetSavePath()))
@@ -36,7 +34,7 @@ namespace LethalProgression.Saving
                 Directory.CreateDirectory(GetSavePath());
             }
 
-            File.WriteAllText(GetSavePath() + "shared.json", JsonConvert.SerializeObject(new SaveSharedData(xp, level, quota)));
+            File.WriteAllText($"{GetSavePath()}shared.json", JsonConvert.SerializeObject(new SaveSharedData(xp, level, quota)));
         }
 
         public static void DeleteSave(int _saveFileSlot)
@@ -45,25 +43,25 @@ namespace LethalProgression.Saving
             // Delete entire folder
             if (Directory.Exists(GetSavePath()))
             {
-                Directory.Delete(Application.persistentDataPath + "/lethalprogression/save" + (saveFileSlot + 1), true);
+                Directory.Delete($"{Application.persistentDataPath}/lethalprogression/save{saveFileSlot + 1}", true);
             }
         }
 
         public static string GetSavePath()
         {
-            return Application.persistentDataPath + "/lethalprogression/save" + (saveFileSlot + 1) + "/";
+            return $"{Application.persistentDataPath}/lethalprogression/save{saveFileSlot + 1}/";
         }
 
         public static string Load(ulong steamId)
         {
             saveFileSlot = GameNetworkManager.Instance.saveFileNum;
 
-            if (!File.Exists(GetSavePath() + steamId + ".json"))
+            if (!File.Exists($"{GetSavePath()}{steamId}.json"))
             {
                 return null;
             }
 
-            string json = File.ReadAllText(GetSavePath() + steamId + ".json");
+            string json = File.ReadAllText($"{GetSavePath()}{steamId}.json");
 
             return json;
         }
@@ -72,13 +70,13 @@ namespace LethalProgression.Saving
         {
             saveFileSlot = GameNetworkManager.Instance.saveFileNum;
 
-            if (!File.Exists(GetSavePath() + "shared.json"))
+            if (!File.Exists($"{GetSavePath()}shared.json"))
             {
                 LethalPlugin.Log.LogInfo("Shared file doesn't exist");
                 return null;
             }
 
-            string json = File.ReadAllText(GetSavePath() + "shared.json");
+            string json = File.ReadAllText($"{GetSavePath()}shared.json");
             LethalPlugin.Log.LogInfo("Shared file exists");
             return JsonConvert.DeserializeObject<SaveSharedData>(json);
         }
