+++ LethalProgression/Skills/SprintSpeed.cs
@@ -1,12 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Reflection.Emit;
-using System.Text;
-using BepInEx.Bootstrap;
 using GameNetcodeStuff;
 using HarmonyLib;
-using BepInEx.Configuration;
 
 namespace LethalProgression.Skills
 {
@@ -16,11 +12,11 @@ namespace LethalProgression.Skills
         public static float sprintSpeed = 2.25f;
         public static void SprintSpeedUpdate(int updatedValue, int newStamina)
         {
-            if (!LP_NetworkManager.xpInstance.skillList.IsSkillListValid())
+            SkillList skillList = LP_NetworkManager.xpInstance.skillList;
+            if (skillList.IsSkillListValid() || skillList.IsSkillValid(UpgradeType.SprintSpeed))
+            {
                 return;
-
-            if (!LP_NetworkManager.xpInstance.skillList.IsSkillValid(UpgradeType.SprintSpeed))
-                return;
+            }
 
             Skill skill = LP_NetworkManager.xpInstance.skillList.skills[UpgradeType.SprintSpeed];
             PlayerControllerB localPlayer = GameNetworkManager.Instance.localPlayerController;
@@ -34,13 +30,15 @@ namespace LethalProgression.Skills
             return sprintSpeed;
         }
 
-        [HarmonyPatch(typeof(PlayerControllerB), "Update")]
+        [HarmonyPatch(typeof(PlayerControllerB), nameof(PlayerControllerB.Update))]
         [HarmonyTranspiler]
         private static IEnumerable<CodeInstruction> Update_Transpiler(IEnumerable<CodeInstruction> instructions)
         {
             // Don't run if mike's tweaks is installed
             if (LethalPlugin.MikesTweaks)
+            {
                 return instructions;
+            }
 
             var codes = new List<CodeInstruction>(instructions);
