+++ LethalProgression/Skills/Strength.cs
@@ -1,10 +1,8 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Linq;
-using System.Text;
 using GameNetcodeStuff;
 using HarmonyLib;
-using Unity.Netcode;
 
 namespace LethalProgression.Skills
 {
@@ -13,19 +11,17 @@ namespace LethalProgression.Skills
     {
         public static void StrengthUpdate(int change = 0, int newLevel = 0)
         {
-            if (!LP_NetworkManager.xpInstance.skillList.IsSkillListValid())
+            SkillList skillList = LP_NetworkManager.xpInstance.skillList;
+            if (!skillList.IsSkillListValid() || !skillList.IsSkillValid(UpgradeType.Strength)
+                || skillList.skills[UpgradeType.Strength].GetLevel() == 0)
+            {
                 return;
-
-            if (!LP_NetworkManager.xpInstance.skillList.IsSkillValid(UpgradeType.Strength))
-                return;
-
-            if (LP_NetworkManager.xpInstance.skillList.skills[UpgradeType.Strength].GetLevel() == 0)
-                return;
+            }
 
             PlayerControllerB player = GameNetworkManager.Instance.localPlayerController;
 
             // Get every object in our hands.
-            List<GrabbableObject> objects = player.ItemSlots.ToList<GrabbableObject>();
+            List<GrabbableObject> objects = player.ItemSlots.ToList();
 
             LethalPlugin.Log.LogDebug($"Carry weight was {player.carryWeight}");
 
@@ -36,7 +32,10 @@ namespace LethalProgression.Skills
             float newCarryWeight = 0f;
             foreach (GrabbableObject obj in objects)
             {
-                if (obj == null) continue;
+                if (obj == null)
+                {
+                    continue;
+                }
 
                 float oldItemWeight = obj.itemProperties.weight - 1f;
                 oldItemWeight *= (1 - multiplier);
@@ -53,7 +52,7 @@ namespace LethalProgression.Skills
         }
 
         [HarmonyPostfix]
-        [HarmonyPatch(typeof(PlayerControllerB), "GrabObjectClientRpc")]
+        [HarmonyPatch(typeof(PlayerControllerB), nameof(PlayerControllerB.GrabObjectClientRpc))]
         private static void UpdateObjects()
         {
             // have to recalculate, grabobject runs too many times.. somewhat inefficient
@@ -61,11 +60,13 @@ namespace LethalProgression.Skills
         }
 
         [HarmonyPostfix]
-        [HarmonyPatch(typeof(GrabbableObject), "DiscardItem")]
+        [HarmonyPatch(typeof(GrabbableObject), nameof(GrabbableObject.DiscardItem))]
         private static void UpdateByDiscard(GrabbableObject __instance)
         {
             if (__instance.IsOwner)
+            {
                 StrengthUpdate();
+            }
         }
     }
 }
