+++ LethalProgression/UpgradeType.cs
@@ -1,0 +1,15 @@
+﻿namespace LethalProgression.Skills
+{
+    public enum UpgradeType
+    {
+        HPRegen,
+        Stamina,
+        Battery,
+        HandSlot,
+        Value,
+        Oxygen,
+        JumpHeight,
+        SprintSpeed,
+        Strength
+    }
+}
